feat: refactored directory and implemented loading status when switching between tabs in my movies tab
